moved data_extraction to data_extractor
