+++ src/rail/plotting/project_dataset_holder.py
@@ -6,7 +6,7 @@ from ceci.config import StageParameter
 
 from rail.projects import RailProject
 from .dataset_holder import RailDatasetHolder
-from .data_extraction import RailProjectDataExtractor
+from .data_extractor import RailProjectDataExtractor
 from .dataset_factory import RailDatasetFactory
